+++ backend/team6/src/main/java/programmers/team6/domain/vacation/rule/MonthlyVacationRule.java
@@ -3,6 +3,7 @@ package programmers.team6.domain.vacation.rule;
 import lombok.RequiredArgsConstructor;
 import programmers.team6.domain.vacation.rule.vacationgranteligiblities.VacationGrantEligibility;
 import programmers.team6.domain.vacation.rule.vacationgranteligiblities.VacationGrantInfo;
+import programmers.team6.global.entity.Positive;
 
 @RequiredArgsConstructor
 public final class MonthlyVacationRule {
