+++ backend/team6/src/main/java/programmers/team6/global/entity/Positive.java
@@ -1,8 +1,9 @@
-package programmers.team6.domain.vacation.rule;
+package programmers.team6.global.entity;
 
 public final class Positive {
 
 	public static final int MIN_VALUE = 0;
+	public static final String NEGATIVE_VALUE_ERROR_MESSAGE = "해당 값을 음수가 될 수 없습니다.";
 
 	private final int value;
 
@@ -12,7 +13,7 @@ public final class Positive {
 
 	private static int requirePositive(int value) {
 		if (value < MIN_VALUE) {
-			throw new IllegalArgumentException("해당 값을 음수가 될 수 없습니다.");
+			throw new IllegalArgumentException(NEGATIVE_VALUE_ERROR_MESSAGE);
 		}
 		return value;
 	}
