+++ backend/team6/src/test/java/programmers/team6/domain/vacation/rule/AnnualVacationRuleTest.java
@@ -11,6 +11,7 @@ import org.junit.jupiter.params.provider.CsvSource;
 
 import programmers.team6.domain.vacation.rule.vacationgranteligiblities.VacationGrantEligibility;
 import programmers.team6.domain.vacation.rule.vacationgranteligiblities.VacationGrantInfo;
+import programmers.team6.global.entity.Positive;
 
 class AnnualVacationRuleTest {
